Improve networking, reliability, and DB persistence
Multiple fixes and hardening across networking, reliability, replication, interpolation, server connection management, logging, and persistence.

Highlights:
- Networking: packet header expanded (16->18 bytes) and ack bitfield 16->32 bits; PacketHeader validation added; payloadSize cross-checked against actual packet length; packet overflow logged and dropped; ack/retransmit logic updated with configurable retransmit delay; pending reliable queue capped (256); getaddrinfo DNS usage noted in docs.
- Reliability: trackReliable now records timestamps, accepts currentTime, pending packet cap enforced, build/process ack signatures use uint32 ackBits, retransmit window widened to 32 bits.
- ConnectionManager: NetAddress hashing + address->client map for O(1) lookup, maxClients cap with set/get and rejection on overflow, removeClient cleans address map.
- Replication: switched from lastAckedState -> lastSentState for diffs, updates and erases adjusted accordingly.
- NetClient/NetServer: cache lastPollTime for RTT-based retransmit calculations and pass to trackReliable; server validates payload sizes before decrypt.
- Packet/protocol: statusEffectMask widened to 32-bit in SvEntityUpdateMsg and read/write adjusted.
- Interpolation: initEntity to seed positions and getInterpolatedPosition reports validity to avoid flash-to-origin.
- Logging/byte stream: LogViewer mutex lock to prevent races; ByteWriter logs buffer overflow warning.
- Animator: guard against frameCount==0 to prevent div-by-zero.
- Docs & DB/persistence notes: WAL fflush on append and replay behavior documented; DbPool/Guard improvements (operator bool and wrap), many production hardening items documented (rate limiter per-account preservation, max client cap, atomic character save, full item metadata in bank, bounty gold deduction, etc.).

These changes aim to close multiple reliability, security, and data-integrity issues and improve server robustness under load.

+++ tests/test_net_integration.cpp
@@ -58,7 +58,7 @@ TEST_CASE("Net Integration: reliable connect handshake via loopback") {
     REQUIRE(client != nullptr);
 
     // Server sends ConnectAccept
-    uint16_t ack; uint16_t ackBits;
+    uint16_t ack; uint32_t ackBits;
     serverRel.buildAckFields(ack, ackBits);
 
     uint8_t respBuf[MAX_PACKET_SIZE];
